fix: form component validation
