--- src/app/shared/formComponent/form.component.html
@@ -1,64 +1,0 @@
-<div class="addFieldWrapper" *ngIf="addFlag">
-  <p>{{ formRequestFields.title }}</p>
-  <form class="form" [formGroup]="profileForm" (ngSubmit)="omSubmit()">
-    <label class="form__label"
-      >{{ formRequestFields.firstRow }}:
-      <input
-        class="form__input_needValidationField"
-        type="text"
-        placeholder="{{ formRequestFields.firstRow | lowercase }}"
-        formControlName="{{ formRequestFields.firstRow }}"
-    /></label>
-    <div
-      class="form__input_invalidFieldSignature"
-      *ngIf="
-        profileForm.controls[formRequestFields.firstRow].invalid &&
-        profileForm.controls[formRequestFields.firstRow].touched
-      "
-    >
-      Undefined or uncorrect name field.
-    </div>
-    <br />
-    <label class="form__label"
-      >{{ formRequestFields.secondRow }}:
-      <input
-        class="form__input_needValidationField"
-        type="text"
-        placeholder="{{ formRequestFields.secondRow | lowercase }}"
-        formControlName="{{ formRequestFields.secondRow }}"
-    /></label>
-    <div
-      class="form__input_invalidFieldSignature"
-      *ngIf="
-        profileForm.controls[formRequestFields.secondRow].invalid &&
-        profileForm.controls[formRequestFields.secondRow].touched
-      "
-    >
-      Undefined or uncorrect Lastname field.
-    </div>
-    <br />
-    <label class="form__label"
-      >{{ formRequestFields.thirdRow }}:
-      <input
-        type="text"
-        placeholder="{{ formRequestFields.thirdRow | lowercase }}"
-        formControlName="{{ formRequestFields.thirdRow }}"
-    /></label>
-    <br />
-    <label class="form__label"
-      >{{ formRequestFields.fourthRow }}:
-      <textarea
-        placeholder="{{ formRequestFields.fourthRow | lowercase }}"
-        formControlName="{{ formRequestFields.fourthRow }}"
-      ></textarea
-    ></label>
-    <br />
-    <button
-      class="form__button_submit"
-      type="submit"
-      [disabled]="!profileForm.valid"
-    >
-      Add
-    </button>
-  </form>
-</div>
