--- src/app/shared/formComponent/form.component.less
@@ -1,41 +1,0 @@
-@import "../../common/constants/colors.less";
-
-.addFieldWrapper {
-  margin: 0 5%;
-  max-width: 350px;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  padding: 15px;
-  border: 3px solid @formBorderColor;
-  border-radius: 5px;
-  &__label {
-
-  }
-  &__button_submit {
-    width: 80px;
-    cursor: pointer;
-    padding: 5px;
-    border-radius: 5px;
-    color: @mainEmblemColor;
-    font-size: 1.7vh;
-    font-weight: 900;
-    border: none;
-  }
-  &__input_needValidationField.ng-touched.ng-invalid {
-    background-color: lighten(@inputInvalidFieldColor, 40%);
-    border: solid @inputInvalidFieldColor 2px;
-  }
-  &__input_needValidationField.ng-touched.ng-valid {
-    background-color: lighten(@inputValidFieldColor, 70%);
-    border: solid @inputValidFieldColor 2px;
-  }
-  &__input_invalidFieldSignature {
-    color:darken(@textColor, 30%);
-    font-style: italic;
-    padding-left: 5%;
-    margin-top: 5px;
-  }
-}
