+++ src/app/shared/shared.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { CommonModule } from "@angular/common";
 
-import { FormComponent } from "./formComponent/form.component";
+import { FormComponent } from "./form/form";
 import { ReactiveFormsModule } from "@angular/forms";
 
 @NgModule({
